requirements.txt
asgiref==3.8.1
Django==5.1.3
django-localflavor==4.0
django-polymorphic==3.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.0
python-stdnum==1.20
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.2
